+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [1.0.3]
+### Changed
+- The project no longer relies on gen-class, instead opting to have a JVM-only dependency.
+
 ### Fixed
 - The `:no-error` clause in `handler-case` was called even on an error